Second in Project Euler Jan 13th start 1:30 end 3:10
I worked for more than 60 minutes and was still not able to finish the second challenge. What I have done so far was make an input variable that leads to an if statement saying if the number is greater than 2 follow the rest. In the rest, I had some help from Nick to set up a way to isolate the two numbers. After isolating them we can add the two numbers together shoving them in a list that holds all the even number answers. Sadly the last part did not work because the return ls on line 25 has a syntax error.