third one in Project Euler 1/17/2022 7:20 pm to 9:36 pm
Okay so first of all this was hard and I spent way too much time on this which is not good at all. Funny thing or sad thing I found almost the exact code from someone else when looking for inspiration after getting assistance. Anyway enough of that, I made a code that has two loops the first one goes until the prime factor is about to be larger than the square root of the number because the prime factor is always less than the square root. The second loop or inside the loop sees if the number dividing the original number is divisible going through that process until unable to do so any longer leaving only the greatest prime factor.