#Fifth one in project Euler 1/17/2022 9:47 to 10:47
I did skip the fourth one it was a nightmare for me so I am now stuck on the fifth problem having way more luck than with the fourth one. So far in project five, I was able to set up code that would divide 2520 by 1 through 10 filterings all the odd out leaving the even ones that are positive. On the code, for this part, I have 1 to 20 because I am now trying to solve the problem by applying the same thing for 2520 but for all numbers below that one to see if I can find the smallest. The only issue with all of that is I don't have any way to tell if the number that goes in instead of 2520 is the smallest possible for 1 to 20 to divide. Oh well, back to the grind and I am happy to report I have an idea for what to do to figure out how to find the smallest number I will let you know in the next update if it worked or not.