#Sixth one in project Euler 1/17/2022 11:47 to 12:47
This project question wasn't so bad because I could look up the equations and plug them in basically which made this problem easy. In my code, I have a lot of unnecessary parts like on line 79 the reason I had those was I was prepared for a lot of complexity trying to make each part as clear as possible. I also didn't need lines 76 to 78 I could have just kept the range but again overestimating the problem made me have all that fluff which I decided to leave since it works. The hardest parts were understanding the equations to know when to put them in the code to complete the project finding sum square differences with natural numbers. Also, I am still a little confused with line 82 where I had square_of_sum * square_of_sum which I saw an equation for when looking at the equation to understand when to use it but it works so I don't care anymore!