Seventh one in the project Euler 1/18/2022 12:47 to 1:48
I don't know if because I am tired but I am very lost on this question I thought I did everything right sort of. The only thing I am missing I believe is more math to find prime numbers even though I tried to set up a code that would constantly just display prime numbers or not which I would filter out. The issue is that my code is not filtering them which I thought I did at line 97 but it's not working or I have to do more with line 100 in getting them out of the list. At least what did go well is my flag variable which I made to be a bool that can help me filter non-prime numbers which it did just I couldn't get rid of the primes is what the issue was. Also, I am really happy with the range that I set up it got a lot of prime numbers but I am wondering now that if I didn't have it I would stop skipping lines allowing for my filter to work better but that is for next time I work on this project.